+++ assets/src/backends/rendering/adapter-base.mjs
@@ -1,4 +1,9 @@
 class RendererAdapter {
+
+    createObject() {}
+    createGroup() {}
+    createCamera() {}
+
 }
 
 export default RendererAdapter;