+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Warning; this changelog is just rough notes and may not be comprehensive.
 - Restructured to be fully modular
 - Now using a static index.html
 - Electron is now primary
+- Split Project into three parts (project, flavor, and renderer). Was that a good idea?
 
 ---
 Changelog started